+++ src/features/mechs/MechsList/MechsList.jsx
@@ -1,22 +1,44 @@
 import React, {Component} from "react";
+import {connect} from "react-redux";
 import {Table} from "semantic-ui-react";
 
 import MechsListHeader from "./MechsListHeader";
 import MechsListRow from "./MechsListRow";
 
+import schema from "app/schema";
+
+import {selectMech} from "../mechsActions";
+import {selectCurrentMech} from "../mechSelectors";
 
 
-export default class MechsList extends Component {
+const mapState = (state) => {
+    const session = schema.from(state.entities);
+    const {Mech} = session;
+
+    const mechs = Mech.all().withModels.map(mechModel => mechModel.getId());
+
+    const currentMech = selectCurrentMech(state);
+
+    return {mechs, currentMech}
+}
+
+const actions = {
+    selectMech,
+};
+
+
+
+export class MechsList extends Component {
 
     render() {
-        const {mechs = [], onMechClicked, currentMech} = this.props;
+        const {mechs = [], currentMech, selectMech} = this.props;
 
-        const mechRows = mechs.map(mech => (
+        const mechRows = mechs.map(mechID => (
             <MechsListRow
-                mech={mech}
-                key={mech.id}
-                onMechClicked={onMechClicked}
-                selected={mech.id === currentMech}
+                mechID={mechID}
+                key={mechID}
+                onMechClicked={selectMech}
+                selected={mechID === currentMech}
             />
         ));
 
@@ -31,3 +53,5 @@ export default class MechsList extends Component {
         );
     }
 }
+
+export default connect(mapState, actions)(MechsList);
