+++ src/features/mechs/MechsList/MechsListRow.jsx
@@ -1,10 +1,39 @@
 import React from "react";
+import {connect} from "react-redux";
 import {Table} from "semantic-ui-react";
 
 import {getWeightClass} from "../mechSelectors";
 
 
-const MechsListRow = ({mech, onMechClicked, selected}) => {
+import schema from "app/schema";
+
+const mapState = (state, ownProps) => {
+    const session = schema.from(state.entities);
+    const {Mech} = session;
+
+    let mech;
+
+    if(Mech.hasId(ownProps.mechID)) {
+        const mechModel = Mech.withId(ownProps.mechID);
+
+        mech = {
+            // Copy the data from the plain JS object
+            ...mechModel.ref,
+            // Provide a default empty object for the relation
+            mechType : {},
+        };
+
+        if(mechModel.type) {
+            // Replace the default object with a copy of the relation's data
+            mech.mechType = {...mechModel.type.ref};
+        }
+    }
+
+    return {mech};
+}
+
+
+const MechsListRow = ({mech = {}, onMechClicked, selected}) => {
     const {
         id = null,
         type = "",
@@ -40,4 +69,4 @@ const MechsListRow = ({mech, onMechClicked, selected}) => {
 }
 
 
-export default MechsListRow;
+export default connect(mapState)(MechsListRow);
